hello_world/requirements.txt
requests
psycopg2-binary
coreapi
regex
python-csv
urllib3
boto3
elementpath
